web/requirements.txt
fastapi
kollavarsham
pillow
pytz
python-multipart
uvicorn
